--- src/services/redis.ts
@@ -1,7 +1,0 @@
-import { Redis } from "ioredis";
-
-export const OrganizationCache = new Redis({
-    host:'localhost',
-    port: 6379
-});
-
